+++ packages/redux-utils/src/auth/register/registerThunks.ts
@@ -1,8 +1,8 @@
 import { createAsyncThunk } from '@reduxjs/toolkit';
-import { Account, AuthAPI, Error, Response } from '../types';
+import { Account, AuthAPI, Error, Payload } from '../types';
 
 const createRegisterThunks = (api: AuthAPI) => {
-  const registerUser = createAsyncThunk<Response, Account, { rejectValue: Error }>(
+  const registerUser = createAsyncThunk<Payload, Account, { rejectValue: Error }>(
     'auth/session/register',
     async (account: Account, thunkApi) => {
       try {
@@ -11,8 +11,11 @@ const createRegisterThunks = (api: AuthAPI) => {
         if (response.status >= 400) {
           return thunkApi.rejectWithValue(response.error as Error);
         }
+        if (!response.data) {
+          return thunkApi.rejectWithValue({ message: 'No register response data' });
+        }
 
-        return response;
+        return response.data;
       } catch (ex) {
         return thunkApi.rejectWithValue({ message: 'Could not register a new account' });
       }