+++ packages/redux-utils/src/auth/requestPassword/requestPasswordThunks.ts
@@ -1,8 +1,8 @@
 import { createAsyncThunk } from '@reduxjs/toolkit';
-import { AuthAPI, Error, Response } from '../types';
+import { AuthAPI, Error, Payload } from '../types';
 
 const createRequestPasswordThunks = (api: AuthAPI) => {
-  const requestPassword = createAsyncThunk<Response, { email: string }, { rejectValue: Error }>(
+  const requestPassword = createAsyncThunk<Payload, { email: string }, { rejectValue: Error }>(
     'auth/password/request',
     async ({ email }, thunkApi) => {
       try {
@@ -11,8 +11,11 @@ const createRequestPasswordThunks = (api: AuthAPI) => {
         if (response.status >= 400) {
           return thunkApi.rejectWithValue(response.error as Error);
         }
+        if (!response.data) {
+          return thunkApi.rejectWithValue({ message: 'No request password response data' });
+        }
 
-        return response;
+        return response.data;
       } catch (ex) {
         return thunkApi.rejectWithValue({ message: 'Could not request a new password' });
       }