+++ packages/redux-utils/src/auth/resetPassword/resetPasswordThunks.ts
@@ -1,8 +1,8 @@
 import { createAsyncThunk } from '@reduxjs/toolkit';
-import { AuthAPI, Error, Response } from '../types';
+import { AuthAPI, Error, Payload } from '../types';
 
 const createResetPasswordThunks = (api: AuthAPI) => {
-  const resetPassword = createAsyncThunk<Response, { token: string; newPassword: string }, { rejectValue: Error }>(
+  const resetPassword = createAsyncThunk<Payload, { token: string; newPassword: string }, { rejectValue: Error }>(
     'auth/password/reset',
     async ({ token, newPassword }, thunkApi) => {
       try {
@@ -11,8 +11,11 @@ const createResetPasswordThunks = (api: AuthAPI) => {
         if (response.status >= 400) {
           return thunkApi.rejectWithValue(response.error as Error);
         }
+        if (!response.data) {
+          return thunkApi.rejectWithValue({ message: 'No reset password response data' });
+        }
 
-        return response;
+        return response.data;
       } catch (ex) {
         return thunkApi.rejectWithValue({ message: 'Could not reset the password' });
       }