+++ packages/redux-utils/src/auth/session/sessionSelectors.ts
@@ -6,8 +6,9 @@ const selectSelf = (state: RootState) => state.auth.session;
 
 const selectUser = createSelector(selectSelf, (state: SessionState) => state.user);
 const selectUserProfile = createSelector(selectSelf, (state: SessionState) => state.profile);
-const selectAccessToken = createSelector(selectSelf, (state: SessionState) => state.token);
+const selectAccessToken = createSelector(selectSelf, (state: SessionState) => state.accessToken);
+const selectRefreshToken = createSelector(selectSelf, (state: SessionState) => state.refreshToken);
 const selectStatus = createSelector(selectSelf, (state: SessionState) => state.status);
 const selectError = createSelector(selectSelf, (state: SessionState) => state.error);
 
-export { selectAccessToken, selectUser, selectUserProfile, selectStatus, selectError };
+export { selectAccessToken, selectRefreshToken, selectUser, selectUserProfile, selectStatus, selectError };